requirements.txt
langchain==1.2.2
langchain-openai==1.1.7
langchain-huggingface==1.2.0
langchain-google-genai==4.1.3
python-dotenv
openai
transformers
huggingface-hub
datasets
accelerate
chromadb
ipykernel
jupyter
langchain-community
sentence-transformers
langchain-chroma